Rename samples to meshes.
