+++ tests/conftest.py
@@ -14,9 +14,9 @@ from geovista.bridge import Transform
 from geovista.crs import WGS84
 from geovista.geometry import coastlines as geometry_coastlines
 from geovista.pantry.data import lam_uk as pantry_lam_uk
-from geovista.pantry.samples import lam_uk as sample_lam_uk
-from geovista.pantry.samples import lfric as sample_lfric
-from geovista.pantry.samples import lfric_sst as sample_lfric_sst
+from geovista.pantry.meshes import lam_uk as sample_lam_uk
+from geovista.pantry.meshes import lfric as sample_lfric
+from geovista.pantry.meshes import lfric_sst as sample_lfric_sst
 
 
 @pytest.fixture()
